move fetchMovieList  from app.js to movieList
